+++ src/Common/LocalRandom.cs
@@ -6,11 +6,17 @@ namespace Rustic;
 /// <summary>Collection of extensions and utility functionality related to <see cref="Random"/> instances.</summary>
 public static class LocalRandom
 {
+#if NET6_0_OR_GREATER
+    /// <summary>Gets the thread-local random pool.</summary>
+    /// <remarks>Returns <see cref="Random.Shared"/> for .NET 6.0 or greater</remarks>
+    public static Random Shared => Random.Shared;
+#else
     [ThreadStatic]
     private static Random? SharedInstance;
 
     /// <summary>Gets the thread-local random pool.</summary>
     public static Random Shared => SharedInstance ??= new();
+#endif
 
     /// <summary>Chooses a element from the collection using the random.</summary>
     /// <typeparam name="T">The type of the value.</typeparam>